shared/requirements.txt
# Shared Configuration Module Requirements

# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Optional dependencies for specific features
pymongo>=4.0.0  # For MongoDB configuration validation
pika>=1.3.0  # For RabbitMQ configuration validation